functions and classes lint and moved to their proper modules
