+++ micro_framework/amqp/connectors.py
@@ -1,41 +1,16 @@
 import json
 import uuid
 
-from kombu import Queue, Connection
+from kombu import Queue
 
 from micro_framework.amqp.amqp_elements import rpc_exchange, \
-    rpc_routing_key, rpc_broadcast_routing_key, default_transport_options
+    rpc_routing_key, rpc_broadcast_routing_key, Publisher
 from micro_framework.amqp.manager import RPCManager
-from micro_framework.amqp.rpc import Publisher
+from micro_framework.amqp.rpc import listen_to_correlation
 from micro_framework.rpc import RPCConnector, RPCConnection
 
 
-class ListenerReceiver:
-    """
-    Implements a method to receive from the Reply EventListener's connection
-    instance with a timeout possibility.
-
-    When the timeout is reached, it raises TimeoutError.
-
-    This class is usually created by the RPCConnection when it sends a new
-    rpc message to the broker and notify the Reply-Listener to send the reply
-    to a specific multiprocess.Connection instance.
-
-    """
-    def __init__(self, connection):
-        self.connection = connection
-
-    def result(self, timeout=None):
-        if timeout is not None:
-            if self.connection.poll(timeout=timeout):
-                return self.connection.recv()
-            raise TimeoutError()
-        while not self.connection.poll(timeout=.5):
-            pass
-        return self.connection.recv()
-
-
-class RPCProducer(RPCConnection):
+class RPCProducer(Publisher, RPCConnection):
     """
     RPCProducer Implements the RPCConnection abstract class methods.
 
@@ -45,12 +20,14 @@ class RPCProducer(RPCConnection):
     It receives a send_to_listener callable, that is responsible to notify
     the EventListener of a new correlation_id to be aware of.
     """
+
     def __init__(
-            self, producer, target_service: str, reply_to_queue: Queue
+            self, amqp_uri, target_service: str, reply_to_queue: Queue,
+            **kwargs
     ):
+        super(RPCProducer, self).__init__(amqp_uri, **kwargs)
         self.target_service = target_service
         self.reply_to_queue = reply_to_queue
-        self.producer = producer
 
     def send(self, payload, *args, **kwargs):
         payload = json.loads(payload)
@@ -64,8 +41,8 @@ class RPCProducer(RPCConnection):
             routing_key = rpc_routing_key(
                 self.target_service, target_id=target_id
             )
-        self.producer.publish(
-            payload, *args, routing_key=routing_key, exchange=exchange,
+        self.publish(
+            payload, routing_key=routing_key, exchange=exchange,
             **kwargs
         )
 
@@ -73,12 +50,15 @@ class RPCProducer(RPCConnection):
         corr_id = str(uuid.uuid4())
         timeout = kwargs.pop("timeout", None)
 
-        from micro_framework.amqp.dependencies import listen_to_correlation
-        # Add this correlation_id to the dict of correlations and their senders
+        # Notify our reply listener of a new correlation_id that will have
+        # ListenerReceiver.
         receiver = listen_to_correlation(corr_id)
         self.send(
             *args, correlation_id=corr_id,
-            reply_to=self.reply_to_queue.routing_key, **kwargs
+            reply_to=self.reply_to_queue.routing_key,
+            mandatory=True,
+
+            **kwargs
         )
         result = receiver.result(timeout=timeout)
         return result
@@ -95,6 +75,7 @@ class AMQPRPCConnector(RPCConnector):
         .reply_to_queue: The queue to which we will tell the RPCServer to
         send the response message.
     """
+
     def __init__(
             self, amqp_uri: str, target_service: str, reply_to_queue: Queue
     ):
@@ -103,7 +84,7 @@ class AMQPRPCConnector(RPCConnector):
         self.reply_to_queue = reply_to_queue
 
     def get_connection(self):
-        return Publisher(
+        return RPCProducer(
             self.amqp_uri, reply_to_queue=self.reply_to_queue,
             target_service=self.target_service
         )
